Fix sign comparison warning Also add override keywords

+++ framework/src/vectorpostprocessors/VolumeHistogram.C
@@ -64,7 +64,7 @@ VolumeHistogram::execute()
     int bin = (_value[_qp] - _min_value) / _deltaV;
 
     // add the volume contributed by the current quadrature point
-    if (bin >= 0 && bin < _nbins)
+    if (bin >= 0 && static_cast<unsigned int>(bin) < _nbins)
       _volume[bin] += computeVolume();
   }
 }
